backend/requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
alabaster==1.0.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
arrow==1.3.0
arxiv==2.2.0
asyncpg==0.30.0
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.5
bibtexparser==1.4.3
bidict==0.23.1
biopython==1.85
black==25.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
chromadb==1.3.5
click==8.2.1
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
Deprecated==1.2.18
distro==1.9.0
docutils==0.21.2
ecdsa==0.19.1
fake-useragent==2.2.0
fastapi==0.116.1
feedparser==6.0.11
filelock==3.19.1
flake8==7.3.0
fonttools==4.59.2
free_proxy==1.1.3
fsspec==2025.9.0
google-api-python-client==2.162.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-storage==2.18.2
google-genai==1.0.0
googlesearch-python==1.3.0
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.9
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
imagesize==1.4.1
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonref==1.1.0
jsonschema==4.25.1
langchain-core==1.1.1
langchain-google-genai==3.2.0
langchain-anthropic==1.2.0
langchain-openai==1.1.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.1
Mako==1.3.10
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mccabe==0.7.0
mcp==1.13.1
mcpadapt==0.1.13
mdurl==0.1.2
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.2
openai==2.8.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.2
passlib==1.7.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
portalocker==3.2.0
posthog==5.4.0
protobuf>=6.33.2
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pymed==0.8.9
pyparsing==3.2.3
pypdf==6.1.0
PySocks==1.7.1
pytest==8.4.2
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.15.1
rapidfuzz==3.14.3
referencing==0.36.2
requests==2.32.5
rich==14.1.0
roman-numerals-py==3.1.0
rpds-py==0.27.1
rsa==4.9.1
scholarly==1.7.11
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
selenium==4.20.0
sgmllib3k==1.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smolagents==1.22.0
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.8
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.47.3
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.17.4
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.3.0
uv==0.8.15
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.3
wsproto==1.2.0
